splitted state into 2 files in the store
