+++ src/components/Counter.js
@@ -1,6 +1,6 @@
 import { useSelector, useDispatch } from "react-redux";
 
-import { counterActions } from '../store/index';
+import { counterActions } from '../store/counter';
 
 import classes from "./Counter.module.css";
